Legacy Commit
This is everything that is now considered legacy code. There is a new
version that's just about ready, and although it shares a lot of
similarities with version 2010.01 and the scattered evolution it's
taken, it is not backwards compatible.

Everything here is available as is, and primarily for reference; what's
coming next on the master branch is where it's at.