Add copy utility
cpstegfs provides an easy way to copy a directory hierarchy from a
traditional file system to a stegfs file system. And by that I mean it
makes sure that as the copy process descends into deeper directories,
the necessary + gets added to allow stegfs to distinguish between files
and folders.

I doesn't yet allow for easy recovery of a hierarchy from stegfs, but it
shouldn't be too long before we can get that implemented.

In actual fact it will copy anything from anywhere to anywhere else,
you'll just end up with directory names that start with +