Rewrite cache and directory handling
Use tree sturcture to represent files/directories in cache. This has the
advantage that it can represent directories as expected! No more
prefixing them with a +

Directories are (still) not stored by stegfs, but they can be
represented while mounted, in a more normal manner. They no longer need
to have a + prefix, but every time the file system is mounted, they do
need to be created. Once a directory is created, any files within it can
be stat'd and should appear (if they're not corrupt).

stegfs now also returns sensible information when statfs is called on
it: total capacity, available capacity, etc.