Initialisation Rewrite
Rewrite stegfs init---command line arguemtn parsing. It's similar to
encrypt, and a lot of code is shared between both stegfs and mkstegfs.
The documentation has been updated, and more importantly, so have some
of the arguments!

When using paramoid mode, the user can select the cipher
algorithms/modes to use---they're not autodetected as there is no
superblock to read this info from.